refactor: rename ApiExceptionBase to ErrorResponse, and moved into package dto/error refactor: rename ApiExceptionBaseDev to DetailedErrorResponseImpl, and moved into package dto/error refactor: rename ApiExceptionBaseProd to ErrorResponseImpl, and moved into package dto/error refactor: all Tests with renamed classes have been refactored refactor: rename GlobalControllerExceptionHandler to GlobalExceptionHandler, and moved into package controller
