+++ src/main/java/pl/byczazagroda/trackexpensesappbackend/dto/error/DetailedErrorResponseImpl.java
@@ -1,35 +1,33 @@
-package pl.byczazagroda.trackexpensesappbackend.exception;
+package pl.byczazagroda.trackexpensesappbackend.dto.error;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
-import java.util.Collections;
 import java.util.List;
 
 /**
- * ApiExceptionBaseDev create business exception message response
- * for profile prod. Exception message has: message and
+ * DetailedErrorResponse create business exception message response for profiles dev and test.
  */
 @Component
-@Profile("prod")
-public class ApiExceptionBaseProd implements ApiExceptionBase {
+@Profile({"dev", "test"})
+public class DetailedErrorResponseImpl implements ErrorResponse {
     @Override
     public String returnExceptionMessage(String message) {
-        return null;
+        return message;
     }
 
     @Override
     public List<String> returnExceptionMessageList(List<String> list) {
-        return Collections.emptyList();
+        return list;
     }
 
     @Override
     public String returnExceptionDescription(String description) {
-        return null;
+        return description;
     }
 
     @Override
     public List<String> returnExceptionDescriptionList(List<String> descriptionList) {
-        return Collections.emptyList();
+        return descriptionList;
     }
 }
