+++ src/main/java/pl/byczazagroda/trackexpensesappbackend/dto/error/ErrorResponseImpl.java
@@ -1,34 +1,35 @@
-package pl.byczazagroda.trackexpensesappbackend.exception;
+package pl.byczazagroda.trackexpensesappbackend.dto.error;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
- * ApiExceptionBaseDev create business exception message response
- * for profiles dev and test.
+ * ErrorResponseImpl create business exception message response
+ * for profile prod. Exception message has: message and
  */
 @Component
-@Profile({"dev", "test"})
-public class ApiExceptionBaseDev implements ApiExceptionBase {
+@Profile("prod")
+public class ErrorResponseImpl implements ErrorResponse {
     @Override
     public String returnExceptionMessage(String message) {
-        return message;
+        return null;
     }
 
     @Override
     public List<String> returnExceptionMessageList(List<String> list) {
-        return list;
+        return Collections.emptyList();
     }
 
     @Override
     public String returnExceptionDescription(String description) {
-        return description;
+        return null;
     }
 
     @Override
     public List<String> returnExceptionDescriptionList(List<String> descriptionList) {
-        return descriptionList;
+        return Collections.emptyList();
     }
 }
