+++ src/test/java/pl/byczazagroda/trackexpensesappbackend/service/FinancialTransactionServiceImplTest.java
@@ -12,7 +12,7 @@ import org.springframework.context.annotation.FilterType;
 import org.springframework.validation.annotation.Validated;
 import pl.byczazagroda.trackexpensesappbackend.controller.FinancialTransactionController;
 import pl.byczazagroda.trackexpensesappbackend.dto.FinancialTransactionDTO;
-import pl.byczazagroda.trackexpensesappbackend.exception.ApiExceptionBase;
+import pl.byczazagroda.trackexpensesappbackend.dto.error.ErrorResponse;
 import pl.byczazagroda.trackexpensesappbackend.exception.AppRuntimeException;
 import pl.byczazagroda.trackexpensesappbackend.exception.ErrorCode;
 import pl.byczazagroda.trackexpensesappbackend.mapper.FinancialTransactionModelMapper;
@@ -42,7 +42,7 @@ class FinancialTransactionServiceImplTest {
     public static final Instant DATE_NOW = Instant.now();
 
     @MockBean
-    private ApiExceptionBase apiExceptionBase;
+    private ErrorResponse errorResponse;
 
     @MockBean
     private FinancialTransactionRepository financialTransactionRepository;
