+++ src/test/java/pl/byczazagroda/trackexpensesappbackend/service/WalletUpdateServiceImplTest.java
@@ -11,7 +11,7 @@ import org.springframework.validation.annotation.Validated;
 import pl.byczazagroda.trackexpensesappbackend.controller.WalletController;
 import pl.byczazagroda.trackexpensesappbackend.dto.UpdateWalletDTO;
 import pl.byczazagroda.trackexpensesappbackend.dto.WalletDTO;
-import pl.byczazagroda.trackexpensesappbackend.exception.ApiExceptionBase;
+import pl.byczazagroda.trackexpensesappbackend.dto.error.ErrorResponse;
 import pl.byczazagroda.trackexpensesappbackend.mapper.WalletModelMapper;
 import pl.byczazagroda.trackexpensesappbackend.model.Wallet;
 import pl.byczazagroda.trackexpensesappbackend.repository.WalletRepository;
@@ -39,7 +39,7 @@ class WalletUpdateServiceImplTest {
     private static final Instant DATE_NOW = Instant.now();
 
     @MockBean
-    private ApiExceptionBase apiExceptionBase;
+    private ErrorResponse errorResponse;
 
     @MockBean
     private WalletRepository walletRepository;
